feat: recompute optimization of llama block
* feat: ctiga block support fused mha and mlp
* fix: adjust pre norm mlp args
* fix: fix fmha install script
* fix: installing fmha_plus should with `no-deps`
* fix: recompute kernel can support fused norm
* style: fix fomat lint
* debug: add control environ
* debug info
* fix: fix fused block usage in ctiga block
* fix: disable recompute as default
* chore: refactor launch.sh
* fix: fix comment
* fix: lint of isort
* style: for black lint
* style: fix format stype with black==22.10.0
* fix: revert

https://code.byted.org/seed/samantha/merge_requests/898

+++ scripts/setup_panther.sh
@@ -1,6 +1,24 @@
-export MARIANA_PANTHER_VERSION=1.7.9.72
-wget http://luban-source.byted.org/repository/scm/lab.speech.panther_arnold_${MARIANA_PANTHER_VERSION}.tar.gz
-tar -zxf lab.speech.panther_arnold_${MARIANA_PANTHER_VERSION}.tar.gz
-sudo pip3 uninstall -y panther-gpu
-pip3 uninstall -y panther-gpu
-pip3 install *torch*/panther_gpu-*.whl
+#!/bin/bash
+# Usage: 
+#   bash scripts/setup_panther.sh 1.7.14.146 or 
+#   OVERRIDE_PANTHER_VERSION=1.7.14.146 bash scripts/setup_panther.sh
+#   This script will prioritize using $1.
+
+OVERRIDE_PANTHER_VERSION=${1:-$OVERRIDE_PANTHER_VERSION}
+if [ -z "${OVERRIDE_PANTHER_VERSION}" ]
+then
+    echo "OVERRIDE_PANTHER_VERSION not set, will not update panther"
+else
+    echo "OVERRIDE_PANTHER_VERSION set, will update panther to ${OVERRIDE_PANTHER_VERSION}"
+    PANTHER_TMP=/tmp/panther_${OVERRIDE_PANTHER_VERSION}
+    rm -fr ${PANTHER_TMP}
+    mkdir -p ${PANTHER_TMP}
+    pushd ${PANTHER_TMP}
+    wget http://luban-source.byted.org/repository/scm/lab.speech.panther_arnold_${OVERRIDE_PANTHER_VERSION}.tar.gz
+    tar -zxf lab.speech.panther_arnold_${OVERRIDE_PANTHER_VERSION}.tar.gz
+    sudo pip3 uninstall -y panther-gpu
+    pip3 uninstall -y panther-gpu
+    sudo -E pip3 install *torch*/panther_gpu-*.whl
+    popd
+    rm -fr ${PANTHER_TMP}
+fi
